+++ docker/.env-non-dev
@@ -46,7 +46,7 @@ REDIS_HOST=redis
 REDIS_PORT=6379
 
 SUPERSET_ENV=production
-SUPERSET_LOAD_EXAMPLES=yes
+SUPERSET_LOAD_EXAMPLES=no
 SUPERSET_SECRET_KEY=TEST_NON_DEV_SECRET
 CYPRESS_CONFIG=false
 SUPERSET_PORT=8088
