+++ docker/pythonpath_dev/superset_config.py
@@ -25,7 +25,6 @@ import os
 
 from celery.schedules import crontab
 from flask_caching.backends.filesystemcache import FileSystemCache
-from custom_security.custom_security_manager import CustomSecurityManager
 
 logger = logging.getLogger()
 
@@ -41,11 +40,6 @@ EXAMPLES_PASSWORD = os.getenv("EXAMPLES_PASSWORD")
 EXAMPLES_HOST = os.getenv("EXAMPLES_HOST")
 EXAMPLES_PORT = os.getenv("EXAMPLES_PORT")
 EXAMPLES_DB = os.getenv("EXAMPLES_DB")
-
-
-CUSTOM_SECURITY_MANAGER = CustomSecurityManager
-
-SECRET_KEY = os.getenv("SUPERSET_SECRET_KEY")
 
 # The SQLAlchemy connection string.
 SQLALCHEMY_DATABASE_URI = (
