--- superset/views/custom_api/api.py
@@ -1,3 +1,0 @@
-"""
-Custom API for management of user with Web Application for external user and role management
-"""
